Added: Navbar & Work Section. Adjusted: word slider & 3D Model
I added the navbar with: "Work, Experience and Skills".
I added the work section. I added three projects with images and short texts.

I reduced the width of the 3D Model to 60% of the screen, so the scrolling becomes more user friendly (UX > UI)
On smaller devices  (< 768px) I completely disabled the OrbitControls, the scrolling function was not intuitive enough.